+++ tests/test_momentum_conversion_forward.py
@@ -5,8 +5,9 @@ from typing import TYPE_CHECKING
 import numpy as np
 import pytest
 import xarray as xr
+from lmfit.models import QuadraticModel
 
-from arpes.fits.fit_models import AffineBroadenedFD, QuadraticModel
+from arpes.fits.fit_models import AffineBroadenedFD
 from arpes.fits.utilities import broadcast_model
 from arpes.utilities.conversion.forward import (
     convert_coordinate_forward,